+++ core/core-cart/src/commonMain/kotlin/com/foodsaver/app/data/dto/CartRequestDto.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 internal data class CartRequestDto(
     val productId: String,
-    val quantity: Int = 1
+    val quantity: Long = 1
 )
